+++ src/in/web/unsupervised.ts
@@ -1,5 +1,0 @@
-import { addCommonUnsupervisedActions } from '@entry/shared/unsupervised';
-
-export function addUnsupervisedActions() {
-    addCommonUnsupervisedActions();
-}