+++ src/tasks.ts
@@ -13,7 +13,7 @@ import { ACTIONS } from '@/config';
 import { notifyOnTaskFinish } from '@events/shells';
 import { addActions, getActions, removeActions } from './rce';
 import { updateActionStatus } from '@events/actions';
-import { RCEContext } from '@context/rce';
+import { RCEContext } from '@ctx/rce';
 
 export const CATEGORY_TASKS = 'Tasks';
 const CATEGORY_REGISTERED_TASKS = 'Registered Tasks';
@@ -25,6 +25,7 @@ export const taskActions = {
         description: 'Terminate the currently running task',
         category: CATEGORY_TASKS,
         handler: handleTerminateTask,
+        // TODO: Auto-switch to task output (if already opened) as preview effect
         cancelEvents: [
             notifyOnTaskFinish,
         ],
@@ -138,6 +139,7 @@ export function reloadTasks() {
             description: task.description,
             category: CATEGORY_REGISTERED_TASKS,
             handler: handleRunTask,
+            // TODO: Auto-switch to task output (if already opened) as preview effect
             promptGenerator: `run the task: ${task.description}`,
             validators: {
                 sync: [checkVirtualWorkspace, checkWorkspaceTrust],